Python/Automatic-Instrumentation/Backend/requirements.txt
Flask==3.1.0
requests==2.26.0
mysql-connector-python==8.0.27
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-mysql
opentelemetry-exporter-jaeger